qcacld-2.0: Fix possible buffer overflow in wma_stats_ext_req
In the function __wlan_hdd_cfg80211_stats_ext_request, data_len
is recieved from vendor command and is passed ultimately to
wma_stats_ext_req. In wma_stats_ext_req, len is calculated as
sum of sizeof(*cmd), WMI_TLV_HDR_SIZE, preq->request_data_len.
The len is of type u_int16_t and adding
sizeof(*cmd) + WMI_TLV_HDR_SIZE will cause a buffer overflow.

Changed the datatype of len to size_t so that it doesn't overflow.

CRs-Fixed: 2350908

+++ drivers/staging/qcacld-2.0/CORE/SERVICES/WMA/wma.c
@@ -27435,7 +27435,7 @@ VOS_STATUS wma_stats_ext_req(void *wda_handle,
 	tp_wma_handle wma = (tp_wma_handle)wda_handle;
 	wmi_req_stats_ext_cmd_fixed_param *cmd;
 	wmi_buf_t buf;
-	u_int16_t len;
+	size_t len;
 	u_int8_t *buf_ptr;
 
 	len = sizeof(*cmd) + WMI_TLV_HDR_SIZE +
